Move utils functions from synth to core.utils
